attribution-mcp/mcp-server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
mcp>=1.0.0
snowflake-connector-python>=3.10.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0
